fix(tests): resolve CI compilation errors in sentry_traversal.rs
Correctly updates `tests/sentry_traversal.rs` to use the new `min_depth` and `max_depth` fields of `PhysicalOp::IndexedTraversal` instead of the removed `depth` field.
Also fixes import issues by correctly restructuring the test file.
This fixes the CI build failure.
