+++ Cargo.toml
@@ -65,7 +65,7 @@ serde_json = { version = "1.0", optional = false }
 
 # ONNX runtime for local model inference
 ort = { version = "2.0.0-rc.0", optional = true }
-tokenizers = { version = "0.22", optional = true }
+tokenizers = { version = "0.15", optional = true }
 num_cpus = { version = "1.16", optional = true }
 
 # Redb for tiered cold storage (pure Rust, ADR-0025)
@@ -83,7 +83,7 @@ rand = "0.8"
 [dev-dependencies]
 criterion = "0.8"
 proptest = "1.4"
-tempfile = "3.26"
+tempfile = "3.8"
 serial_test = "3.4"
 # Tokio for async tests
 tokio = { version = "1.35", features = ["rt-multi-thread", "macros", "test-util"] }
