--- DX_AUDIT_REPORT_ECHO.md
@@ -1,36 +1,0 @@
-# 🗣️ Echo DX Report: The Audit
-
-**Auditor:** Echo (Voice of the User)
-**Subject:** Developer Experience Audit & Fixes
-
-## 🔍 The Walkthrough
-
-I performed a comprehensive audit of the "Getting Started" experience, running examples from the README and verifying their behavior.
-
-### ✅ What Works
-- **Basic Graph Operations:** Copy-pasting the example works flawlessly.
-- **Vector Search:** Works (with minor warnings about unused variables in some contexts).
-- **Narrative Generation:** The feature flag requirement is clearly communicated by the compiler error message ("item is gated behind the `nova` feature").
-- **Sharding:** The example code compiles and runs even without the optional RPC feature (which correctly errors at runtime if used).
-
-### 🚧 Friction Points & Fixes
-
-#### 1. `AletheiaDB` Debug Output (Major)
-**Scenario:** Tried to inspect the database state with `println!("{:?}", db)`.
-**Result:** `error[E0277]: AletheiaDB doesn't implement Debug`.
-**Fix:** Implemented `std::fmt::Debug` for `AletheiaDB` in `src/db/mod.rs`. It now prints useful info like timestamp and durability mode without deadlocking (used `try_lock()`).
-
-#### 2. Unused Variable Warnings (Minor)
-**Scenario:** Running examples often results in `warning: unused variable: ...`.
-**Status:** This is a minor annoyance. The actual example files in `examples/` are mostly clean, but snippets in README might need updates to use the results (e.g., printing them).
-
-#### 3. Type Inference in Docs (Minor)
-**Scenario:** `cold_storage_path("path".into())` vs `cold_storage_path("path")`.
-**Status:** The README uses the correct string literal form which works due to `Into<PathBuf>`. Adding `.into()` manually caused ambiguity errors. The docs are correct.
-
-## 🏁 Conclusion
-
-The library is in good shape. The `Debug` implementation for the main struct significantly improves inspectability for new users.
-
-Signed,
-**Echo** 🗣️
