+++ src/lib.rs
@@ -107,10 +107,9 @@ pub use core::error::{Error, QueryError, Result, StorageError, TemporalError, Tr
 pub use db::{AletheiaDB, VectorIndexBuilder};
 pub use index::{
     AdjacencyIndex, CurrentIndexes, TemporalIndexes,
-    vector::{DistanceMetric, HnswConfig, TemporalVectorConfig},
+    vector::{DistanceMetric, HnswConfig},
 };
 pub use storage::CurrentStorage;
-pub use storage::index_persistence::PersistenceConfig;
 pub use storage::wal::{DurabilityMode, WriteOptions};
 
 // Query planner re-exports (VS-060)
