+++ type_check.rs
@@ -1,0 +1,4 @@
+fn main() {
+    let arr = [1.0, f32::NAN, 2.0];
+    let _: &[f32] = &arr;
+}
